# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreetBite API - Street Vendor Order Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors/{vendorId}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor details
      description: Returns vendor information by ID
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Vendor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    
    put:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor details
      description: Updates vendor information (requires authentication)
      security:
        - bearerAuth: []
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorRequest"
      responses:
        "200":
          description: Vendor updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Vendor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendors/me:
    get:
      operationId: getMyVendorProfile
      tags: [vendors]
      summary: Get authenticated vendor profile
      description: Returns the authenticated vendor's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vendor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  
  # Menu Items
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create new order
      description: Creates a new order from customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      description: Returns order information by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      description: Updates order status (requires vendor authentication)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendors/{vendorId}/orders:
    get:
      operationId: getVendorOrders
      tags: [orders]
      summary: Get vendor orders
      description: Returns all orders for a vendor (requires authentication)
      security:
        - bearerAuth: []
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, preparing, ready, delivered]
      responses:
        "200":
          description: Orders retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendors/{vendorId}/token-counter:
    get:
      operationId: getTokenCounter
      tags: [orders]
      summary: Get current token counter
      description: Returns the next token number for the vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token counter retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenCounterResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    
    # Auth Schemas
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        rating:
          type: number
          format: float
        isOpen:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - phoneNumber
        - name
        - description
        - category
        - rating
        - isOpen
    UpdateVendorRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        isOpen:
          type: boolean
    
    # Menu Item Schemas
...
    OrderItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: integer
        quantity:
          type: integer
      required:
        - menuItemId
        - name
        - price
        - quantity
    Order:
      type: object
      properties:
        id:
          type: integer
        tokenNumber:
          type: integer
        vendorId:
          type: integer
        customerId:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, preparing, ready, delivered]
        total:
          type: integer
          description: Total price in cents/paise
        estimatedTime:
          type: integer
          description: Estimated preparation time in minutes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tokenNumber
        - vendorId
        - items
        - status
        - total
        - estimatedTime
    CreateOrderRequest:
      type: object
      properties:
        vendorId:
          type: integer
        customerId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
          minItems: 1
        total:
          type: integer
        estimatedTime:
          type: integer
      required:
        - vendorId
        - items
        - total
        - estimatedTime
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, delivered]
      required:
        - status
    TokenCounterResponse:
      type: object
      properties:
        nextToken:
          type: integer
      required:
        - nextToken